+++ spec/support/retry/message_formatter.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Retry
   class MessageFormatter
     attr_reader :example
